% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choiceset.R
\name{choiceset}
\alias{choiceset}
\title{Create a choiceset from a data.frame.}
\usage{
choiceset(data, ..., type = c("direct", "ptv", "vote_prop"),
  output = c("choice_list", "choice_matrix", "numeric_list",
  "character"), threshold = NULL, threshold.estimate = c("midpoint",
  "median", "learning"), names = NULL, choice = NULL,
  metric = c("Accuracy", "Kappa", "Sensitivity", "Specificity",
  "Balanced"))
}
\arguments{
\item{data}{The data.frame containing the variables}

\item{...}{OPTIONAL to select variables from data (see dplyr::select)}

\item{type}{Type of the data: direct for a dummy variable, ptv for ptv data and vote_prop for vote probabilities.}

\item{output}{choice_list for a list of character vectors containing the choices of the sets; choice_matrix for a dummy matrix containing 1 for the choices includes in the set and 0 otherwiese; numeric_list for a list containing numbered vectors containing the choices of the sets (1 for the first choiche, 2 for the second, ...);  character for a character vector where each elemebt represents a set connecting the choices with -.}

\item{threshold}{The threshold to use, if left empty 0 is used for vote_prop and threshold.estimate for ptv.}

\item{threshold.estimate}{Only relevant for type ptv. Default is midpoint where the midpoint between min and max of the data is used. Median uses the median of the data and learning uses the function optimal_cutpoint to estimate a cutpoint for every choice.}

\item{names}{Names for the choices, default uses the variable names of data.}

\item{choice}{Only needed for threshold.estimate learning; A variablename as character of the selected choice in data.}

\item{metric}{Only needed for threshold.estimate learning; The metric to use for the learning.}
}
\value{
The choicesets in the form defined with the parameter output.
}
\description{
Create a choiceset from a data.frame.
}
\examples{
cnames <- c("50+","CDA","CU","D66","DENK","FvD","GL","PvdA","PvdD","PVV","SGP","SP","VVD")
vp <- matrix(c(0,10,20,0,0,0,0,0,0,0,60,0,10,
              5,0,0,30,0,0,20,45,0,0,0,0,0,
              0,0,0,0,80,0,0,20,0,0,0,0,0,
              0,0,0,0,0,20,0,0,10,50,0,0,20,
              0,0,0,0,0,0,40,35,0,0,0,25,0),
            nrow = 5, ncol = 13, byrow = TRUE)
vp <- as.data.frame(vp)
choiceset(vp, type = 'ptv', names = cnames)
choiceset(vp, type = 'ptv', output = 'character', names = cnames)
choiceset(vp, type = 'ptv', output = 'character', threshold = 20, names = cnames)
choiceset(vp, type = 'ptv', output = 'choice_matrix', threshold.estimate = 'median', names = cnames)
}
