% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_cutpoint.R
\name{optimal_cutpoint}
\alias{optimal_cutpoint}
\title{Estimates the optimal cutpoints.}
\usage{
optimal_cutpoint(data, ..., choice, metric = c("Accuracy", "Kappa",
  "Sensitivity", "Specificity", "Balanced"), names = NULL)
}
\arguments{
\item{data}{The data.frame containing the variables}

\item{...}{to select variables from data (see dplyr::select)}

\item{choice}{A variablename is character of data wich contains the selected choices}

\item{metric}{The metric to use for estimating the cut points.}

\item{names}{The names of the choices. Uses variable names of data by default.}
}
\value{
A list containing two elements. party.cutpoints contains the choice cutpoints. global.cutpoint contains the cutpoint which is globally the best
}
\description{
Estimates the optimal cutpoints.
}
\examples{
choice <- data.frame(PTV.A = c(0, 10, 5, 4, 7, 1, 2, 0, 0, 10),
                     PTV.B = c(5, 0, 1, 10, 6, 7, 10, 9, 5, 0),
                     PTV.C = c(10, 5, 10, 5, 2, 8, 0, 5, 5, 0),
                     VOTE = c("C", "A", "C", "B", "B", "C", "B","B", "B", "A"))
optimal_cutpoint(choice, PTV.A:PTV.C, choice = "VOTE")
optimal_cutpoint(choice, PTV.A:PTV.C, choice = "VOTE", metric = "Kappa")
}
